refactor(py-client): move URL encoding helpers to utils
`encode_path`/`encode_query` are general on-the-wire URL encoding, not
specific to pre-signed URLs. Move them (and the `_PATH_SAFE`/`_QUERY_SAFE`
sets) from `presign` to `utils` so `_make_url`/`_make_multipart_url` and the
pre-signing code call one shared, neutrally-located helper instead of the
request path depending on the `presign` module.
